Properly enqueue styles, and update page template
Previously the stylesheets were directly linked to in the <head> element
in header.php as a workaround to allow the old and new themes to be used
simultaneously. Changes I made meant that that didn't work any more, so
I've properly enqueued the new styles and unenqueued the old ones.

Also updated page.php template so most pages show the basic two column
layout.